requirements.txt
langgraph>=0.2.0
openai>=1.31.0
pydantic>=2.7.0
python-dotenv>=1.0.1
streamlit>=1.36.0
requests>=2.31.0
pytest>=8.2.0
